Add SEV-ES support
Implement support for remotely attesting SEV-ES guests, using a couple
of VMSA dumps (one from a BP, and the other from an AP) extracted from
vCPUs configured by libkrun on a host running linux-5.14.4.

This change also adds a the "sev-es" flag, along with the "cpus" and
"mem" options to the command line. The first one is used to specify
instruct the guest that is should enable SEV-ES, while the others are
used to inject the resource configuration into the guest.

Signed-off-by: Sergio Lopez <slp@redhat.com>